refactor: consolidate globals into config struct, split interceptors.go
Replace 34+ scattered package-level vars with a single interceptorConfig
struct (defaultConfig). Split 936-line interceptors.go into 8 feature files:

- config.go: interceptorConfig struct, defaults, all Set*() functions
- filter.go: FilterMethods, filterState, FilterMethodsFunc
- server.go: all server interceptors (timeout, logging, tracing, etc.)
- client.go: all client interceptors (hystrix, newrelic, retry)
- chain.go: interceptor chain helpers
- http.go: DoHTTPtoGRPC, NRHttpTracer
- metrics.go: Prometheus metrics setup
- ratelimit.go: rate limiting interceptor

Zero public API changes. All Set*() functions now assign to
defaultConfig fields. resetGlobals() test helper simplified.
