+++ chain.go
@@ -1,0 +1,52 @@
+package interceptors
+
+import (
+	"context"
+
+	"google.golang.org/grpc"
+)
+
+// chainUnaryServer chains multiple unary server interceptors into one.
+func chainUnaryServer(interceptors []grpc.UnaryServerInterceptor) grpc.UnaryServerInterceptor {
+	return func(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
+		chain := handler
+		for i := len(interceptors) - 1; i >= 0; i-- {
+			interceptor := interceptors[i]
+			next := chain
+			chain = func(ctx context.Context, req any) (any, error) {
+				return interceptor(ctx, req, info, next)
+			}
+		}
+		return chain(ctx, req)
+	}
+}
+
+// chainUnaryClient chains multiple unary client interceptors into one.
+func chainUnaryClient(interceptors []grpc.UnaryClientInterceptor) grpc.UnaryClientInterceptor {
+	return func(ctx context.Context, method string, req, reply any, cc *grpc.ClientConn, invoker grpc.UnaryInvoker, opts ...grpc.CallOption) error {
+		chain := invoker
+		for i := len(interceptors) - 1; i >= 0; i-- {
+			interceptor := interceptors[i]
+			next := chain
+			chain = func(ctx context.Context, method string, req, reply any, cc *grpc.ClientConn, opts ...grpc.CallOption) error {
+				return interceptor(ctx, method, req, reply, cc, next, opts...)
+			}
+		}
+		return chain(ctx, method, req, reply, cc, opts...)
+	}
+}
+
+// chainStreamClient chains multiple stream client interceptors into one.
+func chainStreamClient(interceptors []grpc.StreamClientInterceptor) grpc.StreamClientInterceptor {
+	return func(ctx context.Context, desc *grpc.StreamDesc, cc *grpc.ClientConn, method string, streamer grpc.Streamer, opts ...grpc.CallOption) (grpc.ClientStream, error) {
+		chain := streamer
+		for i := len(interceptors) - 1; i >= 0; i-- {
+			interceptor := interceptors[i]
+			next := chain
+			chain = func(ctx context.Context, desc *grpc.StreamDesc, cc *grpc.ClientConn, method string, opts ...grpc.CallOption) (grpc.ClientStream, error) {
+				return interceptor(ctx, desc, cc, method, next, opts...)
+			}
+		}
+		return chain(ctx, desc, cc, method, opts...)
+	}
+}
